fix(PUF-206): move precheck to puffo_agent/__init__.py + bump to 0.8.6
Addresses my own non-blocking observation on the round-1 review:
Python parses an entire module before executing any line, so the
precheck-in-cli.py only protects against unparseable *submodules* of
cli.py, not against cli.py itself becoming unparseable. Moving the
precheck up to puffo_agent/__init__.py makes it fire before cli.py
(or any other puffo_agent.* submodule) is even parsed.

- puffo_agent/__init__.py: precheck function + module-load call.
  Module docstring documents the "stays parseable on 3.6+"
  constraint so the discipline isn't lost.
- portal/cli.py: precheck removed; `import sys` returned to the
  stdlib import block where it belongs.
- test_cli_python_version.py: import from `puffo_agent` instead of
  `puffo_agent.portal.cli`; docstring updated to reflect the new
  placement.

Test suite: test_cli_python_version.py 4/4 pass locally. Full suite
otherwise unchanged. (The 2 pre-existing Windows-only failures in
PUF-217's test_refresh_oneshot_home_env.py also fail on the
unmodified branch — `os.rename` doesn't atomically replace on
Windows, needs `os.replace`. Filed mentally as a follow-up; not
caused by this change.)

Bump to 0.8.6 + changelog entry under the canonical PUF-206 ticket.
