+++ src/puffo_agent/portal/cli.py
@@ -7,28 +7,6 @@ Entry point: the ``puffo-agent`` console script, or
 
 from __future__ import annotations
 
-import sys
-
-
-def _require_python_311() -> None:
-    """Bail with a clear, actionable message on Python <3.11. Runs
-    at module load so submodule imports never get a chance to raise
-    a downstream SyntaxError/ImportError that users mis-attribute."""
-    if sys.version_info < (3, 11):
-        v = sys.version_info
-        sys.stderr.write(
-            f"puffo-agent requires Python >= 3.11; you have "
-            f"{v.major}.{v.minor}.{v.micro}.\n"
-            f"Upgrade via pyenv (`pyenv install 3.12`), Homebrew "
-            f"(`brew install python@3.12`), or python.org/downloads, "
-            f"then re-run.\n"
-        )
-        raise SystemExit(1)
-
-
-_require_python_311()
-
-
 import argparse
 import asyncio
 import io
@@ -36,6 +14,7 @@ import logging
 import os
 import shutil
 import subprocess
+import sys
 import time
 from pathlib import Path
